python-version/requirements.txt
slackclient
python-dotenv
certifi
